added members app; moved everything from index; altered script to use correct model;
