+++ project/f5index/admin.py
@@ -1,6 +1,5 @@
 from django.contrib import admin
-from .models import Member, SupportSubmission
+from .models import SupportSubmission
 
 # Register your models here.
-admin.site.register(Member)
 admin.site.register(SupportSubmission)
