+++ project/f5members/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class F5MembersConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'f5members'
