+++ project/f5members/templates/f5members/contact_support.html
@@ -1,0 +1,25 @@
+{% extends "base.html" %}
+
+{% block loadcss %}
+
+{% endblock %}
+
+{% block title %}
+Contact support
+{% endblock %}
+
+{% block content %}
+<div class="section">
+  <div class="form_wrapper auth_form">
+    <h2>Contact support <i class="fa fa-life-ring" aria-hidden="true"></i></h2>
+    <form method="post" action="{% url 'index:contact_support' %}">
+      {% csrf_token %}
+      {{ form.as_p }}
+      <button class="submit" type="submit">Send</button>
+      <div class="button_link_group">
+        <a target="_blank" href="https://en.wikipedia.org/wiki/Pigeon_post#:~:text=Pigeons%20are%20effective%20as%20messengers,many%20places%20around%20the%20world." class="button_link neutral">Who sends this message?</a>
+      </div>
+    </form>
+  </div>
+</div>
+{% endblock %}
