+++ project/f5members/templates/f5members/login_member.html
@@ -1,0 +1,22 @@
+{% extends "base.html" %}
+
+{% block title %}
+Login account
+{% endblock %}
+
+{% block content %}
+<div class="section">
+  <div class="form_wrapper auth_form">
+    <h2>Login</h2>
+    <form method="post" action="{% url 'index:login_member' %}">
+      {% csrf_token %}
+      {{ form.as_p }}
+      <button class="submit" type="submit">Login</button>
+      <div class="button_link_group">
+        <a href="{% url 'index:create_member' %}" class="button_link positive">Create an account.</a>
+      </div>
+    </form>
+  </div>
+</div>
+
+{% endblock %}
