+++ project/f5members/templates/f5members/verify_sent.html
@@ -1,0 +1,21 @@
+{% extends "base.html" %}
+
+{% block loadcss %}
+
+{% endblock %}
+
+{% block title %}
+Verify account
+{% endblock %}
+
+{% block content %}
+<div class="section">
+  <h2>Verification is easy <i class="fa fa-lock" aria-hidden="true"></i></h2>
+  <p>Just check your email, we sent you something secret.</p>
+
+  <div class="button_link_group">
+    Having trouble?
+    <a href="{% url 'index:contact_support' %}" class="button_link neutral">Contact support <i class="fa fa-envelope-o" aria-hidden="true"></i></a>
+  </div>
+</div>
+{% endblock %}
