+++ project/f5members/templates/f5members/verify_user.html
@@ -1,0 +1,22 @@
+{% extends "base.html" %}
+
+{% block loadcss %}
+
+{% endblock %}
+
+{% block title %}
+Verify account
+{% endblock %}
+
+{% block content %}
+<div class="section">
+  <h2>Told you it was easy. <i class="fa fa-lock" aria-hidden="true"></i></h2>
+  {% for message in messages %}
+    <p>{{ message }}</p>
+  {% endfor %}
+  
+  <div class="button_link_group">
+    <a href="{% url 'index:login_member' %}" class="button_link positive">Login!</a>
+  </div>
+</div>
+{% endblock %}
